+++ lightsim2grid/gridmodel/initGridModel.py
@@ -41,8 +41,9 @@ def init(pp_net):
     - the pandapower grid has dcline
     - the pandapower grid has switch, motor, assymetric loads, etc.
     - the pandapower grid any parrallel "elements" (at least one of the column "parrallel" is not 1)
+    - the bus indexes in pandapower do not start at 0 or are not contiguous (you can check `pp_net.bus.index`)
     - some `g_us_per_km` for some lines are not zero ? TODO not sure if that is still the case !
-    - some `p_mw` for some shunts are not zero
+    - some `p_mw` for some shunts are not zero ? TODO not sure if that is still the case !
 
     if you really need any of the above, please submit a github issue and we will work on their support.